// src/app/generated_services/model/createNotificationDTO.ts
/**
 * Carlson Gracie Backend API
 * API for managing students, belts, graduation requirements, notices, frequencies and user authentication with Keycloak
 *
 * Contact: support@carlsongracie.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { NotificationPriority } from './notificationPriority';
import { NotificationType } from './notificationType';


export interface CreateNotificationDTO { 
    title: string;
    message: string;
    type: NotificationType;
    priority?: NotificationPriority;
    userId?: string | null;
    isActive?: boolean;
    expiresAt?: string | null;
    actionUrl?: string | null;
    metadata?: string | null;
}
export namespace CreateNotificationDTO {
}