// src/app/generated_services/model/notificationPriority.ts
/**
 * Carlson Gracie Backend API
 * API for managing students, belts, graduation requirements, notices, frequencies and user authentication with Keycloak
 *
 * Contact: support@carlsongracie.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type NotificationPriority = 0 | 1 | 2 | 3;

export const NotificationPriority = {
    NUMBER_0: 0 as NotificationPriority,
    NUMBER_1: 1 as NotificationPriority,
    NUMBER_2: 2 as NotificationPriority,
    NUMBER_3: 3 as NotificationPriority
};